+++ dipu/torch_dipu/csrc_dipu/base/basedef.h
@@ -9,12 +9,8 @@
 #define C10_COMPILE_TIME_MAX_DIPUS 16
 
 #define DIPU_DEVICE_TYPE_MACRO XPU
-#define DIPU_AUTOGRAD_DEVICE_TYPE_MACRO AutogradXPU
-
-#define DeviceTypeDIPU  at::DeviceType::XPU
-#define AutocastDIPU  AutocastXPU
-
-#define ADD_NS(RAW_OP) at::RAW_OP
+#define DIPU_AUTOGRAD_DEVICE_TYPE_MACRO C10_CONCATENATE(Autograd, DIPU_DEVICE_TYPE_MACRO)
+#define DIPU_AUTOCAST_DEVICE_TYPE_MACRO C10_CONCATENATE(Autocast, DIPU_DEVICE_TYPE_MACRO)
 
 // to do: abstract a layer which not depend on pytorch
 namespace dipu {
@@ -22,12 +18,12 @@ namespace dipu {
 // XPU is originally intel output-of-tree code https://github.com/intel/intel-extension-for-pytorch ( branch xpu-master )
 // we use this type but PrivateUse1 not to impersonate our DIPU device. because compared with PrivateUse1,
 // XPU has richer support in pytorch trunk and not too much feature in torch to interfere our logic (as XLA).
-const auto DIPU_DEVICE_TYPE = at::DeviceType::XPU;
+const auto DIPU_DEVICE_TYPE = at::DeviceType::DIPU_DEVICE_TYPE_MACRO;
 
-const auto DIPU_DISPATCH_KEY = c10::DispatchKey::XPU;
-const auto DIPU_DISPATCH_AUTOGRAD_KEY = c10::DispatchKey::AutogradXPU;
+const auto DIPU_DISPATCH_KEY = c10::DispatchKey::DIPU_DEVICE_TYPE_MACRO;
+const auto DIPU_DISPATCH_AUTOGRAD_KEY = c10::DispatchKey::DIPU_AUTOGRAD_DEVICE_TYPE_MACRO;
 
-const auto DIPU_Backend_TYPE = c10::Backend::XPU;
+const auto DIPU_Backend_TYPE = c10::Backend::DIPU_DEVICE_TYPE_MACRO;
 
 const auto DICL_BACKEND_NAME = "dicl";
